Split routes into different file
